+++ src/utils/charMap.ts
